debug/requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws_ec2==1.128.0
aws-cdk.aws_iam==1.128.0
